requirements.txt
chainlit~=1.1.402
langchain~=0.2.12
langchain-openai
langchain-ollama
langchain-experimental
langgraph~=0.2.3
psycopg2-binary
numpy~=1.26.4
openai~=1.40.6
python-dotenv